# netlify/functions/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Salas API
  description: API para gerenciar salas
paths:
  /salas:
    get:
      summary: Retorna todas as salas
      responses:
        "200":
          description: Lista de salas
    post:
      summary: Adiciona uma nova sala
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                predio:
                  type: string
                andar:
                  type: integer
                sala:
                  type: integer
                descricao:
                  type: string
      responses:
        "200":
          description: Sala adicionada com sucesso
    delete:
      summary: Deleta todas as salas
      responses:
        "200":
          description: Todas as salas foram deletadas com sucesso
  /salas/{id}:
    get:
      summary: Retorna uma sala por ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID da sala
      responses:
        "200":
          description: Detalhes da sala
    put:
      summary: Atualiza uma sala por ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID da sala
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                predio:
                  type: string
                andar:
                  type: integer
                sala:
                  type: integer
                descricao:
                  type: string
      responses:
        "200":
          description: Sala atualizada com sucesso
    delete:
      summary: Deleta uma sala por ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID da sala
      responses:
        "200":
          description: Sala deletada com sucesso